refactor: Fix circular import in debug agent
- Create backend/src/services/deploymentService.js
- Move updateDeploymentStatus and getLatestDeployment to service
- Update imports in debugAgent.js and buildPipeline.js
- Re-export from routes/deployments.js for compatibility

Closes #97
